Add support for cookies
Modern versions of the firmware require cookies. This request the main
URL, without authorisation, to force setting a cookie and arranges for
further calls to use these cookies.

Support for `wget` was dropped to simplify the implementation and avoid
having to manually parse `Set-Cookie` and setting cookies using `Cookie` headers.